Edited on 2-11-2022

+++ src/components/2.HomePage/Properties.js
@@ -26,9 +26,9 @@ const Properties = ({ propertyData }) => {
                         </span>
                         <br />
                         <span className="text-capitalize">
-                          Ranging from {parseFloat(property.minvalue) / 100000}{" "}
+                          Ranging from {parseInt(property.minvalue) / 100000}{" "}
                           Lac
-                          {/* to {parseFloat(property.maxvalue) / 10000000} Cr */}
+                          {/* to {parseInt(property.maxvalue) / 10000000} Cr */}
                         </span>
                         <br />
                         <div className="mt-3">
